// src/api/model/bankAccountGet.ts
/**
 * Financial Management API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BankAccountGet { 
    /**
     * Id of the created resource
     */
    readonly id: string;
    bankId: string;
    /**
     * ISO 4217 currency code
     */
    currency: string;
    readonly ibanCode?: string;
    readonly accountNumber?: string;
    readonly balance?: number;
    readonly createdAt?: string;
    readonly userProfileId?: string;
}