// src/api/model/transactionGet.ts
/**
 * Financial Management API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TransactionGet { 
    /**
     * Id of the created resource
     */
    readonly id: string;
    amount: number;
    description: string;
    /**
     * Type of the transaction
     */
    type: TransactionGet.TypeEnum;
    /**
     * Fee applied to the transaction
     */
    fee?: number;
    sourceAccountId: string;
    destinationAccountId: string;
    readonly createdAt?: string;
}
export namespace TransactionGet {
    export type TypeEnum = 'Withdraw' | 'Transfer' | 'Deposit';
    export const TypeEnum = {
        Withdraw: 'Withdraw' as TypeEnum,
        Transfer: 'Transfer' as TypeEnum,
        Deposit: 'Deposit' as TypeEnum
    };
}