// src/api/model/transactionsPost.ts
/**
 * Financial Management API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TransactionsPost { 
    amount: number;
    description: string;
    /**
     * Type of the transaction
     */
    type: TransactionsPost.TypeEnum;
    /**
     * Fee applied to the transaction
     */
    fee?: number;
    sourceAccountId: string;
    destinationAccountId: string;
}
export namespace TransactionsPost {
    export type TypeEnum = 'Withdraw' | 'Transfer' | 'Deposit';
    export const TypeEnum = {
        Withdraw: 'Withdraw' as TypeEnum,
        Transfer: 'Transfer' as TypeEnum,
        Deposit: 'Deposit' as TypeEnum
    };
}